Refactor the cron package.
  - Have separate concepts for a crontab and the current config of that
crontab. This let's us separate what a crontab is from what it does a
little better (e.g. a user has a crontab and it has a current config
that tells us what it does. We can separate the details of where that
crontab is stored and how to access it from the commands the crontab
runs).
  - Split these two concepts into separate files.