Simplify some of the pipeline stages.
I'm willing to believe that if I pass a ctx to a function that also
accepts a channel, then it will close that channel when the ctx is done
(especially if I'm the one writing the function). I don't think the
increased complexity is really worth the additional layers of safety of
having all the parts of the pipeline check the ctx themselves, even if
they aren't really doing anything expensive.